refactor(core): move build_gitignore_matcher into tgrep-core::walker
Per review: don't fork the gitignore-loading logic between core and cli. Hoist build_gitignore_matcher into tgrep_core::walker (alongside walk_dir / walk_file_metadata) and re-export ignore::gitignore::Gitignore as tgrep_core::walker::Gitignore so the cli can consume it without a direct ignore-crate dep.

Also fixes a discovery bug spotted by the new test: hidden(true) on the WalkBuilder used to enumerate .gitignore files filters dotfiles, including .gitignore itself, so no rules were ever loaded. Walk with hidden(false) and an explicit filter_entry to skip .git/ instead.

tgrep-cli now: drops the ignore = 0.4 dep, calls tgrep_core::walker::build_gitignore_matcher, and the unit test exercises the shared loader by writing a .gitignore to a TempDir and round-tripping through it.
